fresh dev

+++ database/migrations/2023_08_01_105542_create_plans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('plans', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price')->unsigned();
+            $table->decimal('price')->unsigned();
             $table->string('type');
             $table->integer('count_sessions')->unsigned();
             $table->timestamps();
